pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["sr-pato <felipe.adeildo@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.1"
bcrypt = "^4.1.2"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.1"
flask-session = "^0.6.0"
flask-wtf = "^1.2.1"
requests = "^2.31.0"
pandas = "^2.2.0"
openpyxl = "^3.1.2"
matplotlib = "^3.8.2"


[tool.poetry.group.dev.dependencies]
ipython = "^8.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
